dealbrain/mcp_servers/calendar-mcp/requirements.txt
fastapi
uvicorn[standard]
google-api-python-client
google-auth
google-auth-httplib2
python-dotenv
pydantic
